todoapp_withauth/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["faheemameen <mohammadfaheem514@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
sqlmodel = "^0.0.18"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
psycopg = {extras = ["binary"], version = "^3.1.19"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
